Refactor data generation to separate utility class
Moved data generation functions from InformationSchemaLoaderTest to new DataGen utility class. This improves code modularity and reusability, ensuring cleaner test implementations and easier maintenance.
